requirements.txt
psycopg2-binary==2.9.9
pydicom==2.4.4
pynetdicom==2.1.1
python-dateutil==2.9.0.post0
six==1.16.0
Unidecode==1.3.8
google-cloud-storage==2.19.0
